Set up the class to automatically create the ARIA component
The class as it is requires the end-developer to create an
element with id="speech" and the correct ARIA settings. But
they might not know the right way to do that, and the whole
point of this kind of abstraction is to keep things simple.
So I figured the class should just do it for us. I've tested
these changes on 62.0 (64-bit) and Chrome 68.0.3440.106
(Official Build) (64-bit) using NVDA 2018.2 on Windows 10,
and it worked great with both Web Speech and ARIA.